file: Ignore config file with '.' in name
Uci doesn't support files with '.' in the config name and also have
parsing problems confusing option section with the actual filename.

Example:

hello.apk-new is parsed as hello filename and apk-new section

To correctly handle and prevent any kind of error, skip parsing these
file entirely. This is now needed for APK support as it does generate
config with a .apk-new suffix if a config already exist and the package
is updated.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>